// src/app/models/objectOfLaborMaterial.ts
export class ObjectOfLaborMaterial {
    objectOfLaborMaterialId: string | null = null;
    quantity: number | null = null;
    objectOfLaborId: string | null = null;
    materialId: string | null = null;
    materialName: string | null = null;
    stockQuantity: number | null = null;
    totalCount: number | null = null;
}